requirements.txt
keras 
numpy
tensorflow
protobuf
streamlit
opencv-python
opencv-python-headless
pillowgit